+++ packages/react-instantsearch/src/connectors/hierarchicalMenuLogic.js
